src/ml/trainer/requirements.txt
google-cloud-storage==2.10.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0